refactor(ui): address eleventh self-review round — eliminate formatStepName re-export shim
Migrate remaining callers (action-type, agent-output-review-panel, runs/page)
to import formatStepName directly from @/lib/format. Remove the re-export from
task-utils.ts; the internal usage stays as a bare import.
